Fix #79: Return user on login mutation (#80)
**Changes:**
* Change login mutation to use custom code rather than graphql_jwt.ObtainJSONWebToken

The previous login mutation was directly from the graphql_jwt library as is. I've used mostly the same code as that one (graphql_jwt.JSONWebTokenMutation) and added the user. It wasn't mentioned in the official docs but it's probably the way it's supposed to be used, so shouldn't break anytime soon.